v2

+++ database/migrations/2022_11_12_231917_create_semester_results_table.php
@@ -18,7 +18,7 @@ class CreateSemesterResultsTable extends Migration
             $table->timestamps();
             $table->integer('exam_id');
             $table->integer('student_id');
-            $table->integer('gpa');
+            $table->float('gpa');
             $table->integer('status');
         });
     }
